Replace timeseries API with timeframe
The old timeseries API has been replaced with a new timeframe API, with
a few changes:

- expects source/currencies instead of base/symbols
- returns a quotes key instead of a rates key
- quotes are keyed by the source/target currency pair (e.g. GBPUSD)

The API now requires an access key, which can be set on the
RequestBuilder object, or passed via the ExchangeRate class (to avoid
having to always instantiate a RequestBuilder and a Client).

The free version of the API no longer supports HTTPS, so the default
scheme is set to HTTP with an option to change if you have a paid plan.

This also creates a new ServiceException which is thrown if the API
returns a non-successful response.